validate_model/requirements.txt
torch>=2.0.0
transformers>=4.30.0
Pillow>=10.0.0
tqdm>=4.65.0
google-cloud-storage>=2.10.0
